Add dacs lib and dacs test.

+++ K2Tree/include/base.h
@@ -16,7 +16,7 @@ namespace libk2tree {
     using std::string;
 
     typedef unsigned char uchar;
-    typedef size_t uint;
+    typedef unsigned int uint;
 
 //    const size_t k = 2;
 //    const size_t kL = 2;
